src/requirements.txt
# Logo Recognition System Requirements
# Based on: "Scalable Logo Recognition using Proxies" (IEEE WACV 2019)

# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
Pillow>=8.0.0
tqdm>=4.50.0

# Data processing
opencv-python>=4.5.0
scikit-learn>=0.24.0

# Visualization & logging
matplotlib>=3.3.0
tensorboard>=2.4.0

# Configuration
PyYAML>=5.4.0

# Optional: For faster data loading
# lmdb>=1.0.0

# Optional: For distributed training
# torch-distributed>=0.1.0
